feat: add pricer command for estimating BOM cost
This is a rough first pass at a command line utility for taking a list of URLs
from a bill of materials ("BOM") for individual components. Each URL is a link
to a vendor's website that sells the component. This utility attempts to scrape
the price and shipping cost to calculate a total.